refactor(domain): extract World core and timer logic
- Move World struct and NewWorld to internal/domain/system/world.go
- Integrate TurnTimer directly into World
- Move population logic to internal/domain/system/population.go
- Align unit tests for World and Timer
